+++ js/util.js
@@ -1,0 +1,12 @@
+// Функция генерации случайных чисел от 0 до 100
+
+function randomNumber (a, b) {
+  const lower = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
+  const upper = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+}
+
+randomNumber(0,9);
+
+export {randomNumber};
